Add benchmark to test the speed of CNode::ReceiveMsgBytes
Summary
---

This is a low-level benchmark that basically tests the speed of the
low-level function `CNode::ReceiveMsgBytes`, which is a function that
parses the raw bytes coming in off a socket and packages it into
messages for later processing.

This benchmark may be needed later if we want to optimize this function,
and can already be used in conjunction with !1744 to evaluate the gains
from that MR.

In order to accomplish this benchmark, we add a ~1.5MB file to the
`bench/data` directory which contains raw bytes received as packet-sniffed
from a real network connection.

Test Plan
---

- `ninja all bench_bitcoin && ./src/bench/bench_bitcoin -filter='CNode.*'`