Alternative fix for Qt GUI RPC Console using an "ephemeral" config object
Summary
---

This MR is similar in purpose to !1778 but is much shorter and less
extensive a set of changes.

It works by not requiring such extensive modification to the RPC API as
!1778 does.  It fixes bug #508 in an alternate way: we simply define
`GetMutableConfig()` as an alternative to `GetConfig()` which returns a
non-`const` `Config &` object.  We then pass this down to
`Node::executeRpc()`.

An advantage to this approach is this: Preserving the ability to work
with a non-`const` `Config` object may be advantageous and/or useful in
the future should we decide to add the ability to do hot-modification of
config params to the node via RPC (provided we add locks to ensure
thread safety).

Test Plan
---

See issue #508 for how to reproduce the bug in question. Alternatively:

- Use current master.
- Start a testnet node with some custom `-excessiveblocksize=xxx` and
  `-blockmaxsize=yyy` and observe the HTTP RPC from the command-line
  shows the right values for the RPC methods `getexcessiveblock` and
  `getmininginfo`, but the Qt GUI debug console RPC for those commands
  does not.
- Try the above again with this commit and see the problem is fixed.