Bump node expiry to May 15, 2025, introduce tentative "upgrade11"
Summary
---

After finishing upgrade10 work, and in preparation for our next release
that will include this upgrade, we need to bump the node expiry to be
for upgrade11.

This MR introduces the tentative upgrade11 for May 15, 2025 (Nov. 15,
2024 for chipnet). And uses it where appropriate (node expiry, mainly).

Additionally a unit test was added to test the logic added to
activation.cpp for IsUpgrade11Enabled().

This MR should only be merged after we merge all the MRs related to
Upgrade 10 (namely: !1782 and its prereqs).

What's missing from this MR that needs to be done sometime after May 15,
2024: Bump the `check-upgrade-activated` CMake targets to use upgrade11,
not upgrade 10. Currently it's still going to be useful to have these
targets vs. upgrade 10, not 11. But after upgrade 10 activates, we will
need to bump these targets.

Test Plan
---

- Check the UTC times introduced into `chainparams.cpp` are correct.
- `ninja all check-all`
- `ninja check-upgrade-activated-extended`