CI: Save more sanitizer-undefined artifacts on failure
Summary
---

The sanitizer-undefined CI job was not saving all of the logs on
failure. This CI job is optional and may fail, but it's still useful to
see all of the bitcoind logs and other things when trying to
troubleshoot spurious or not-so-spurious failures.

This MR adds such a facility.

Test Plan
---

Verify that the artifacts.zip for the `sanitizer-undefined` CI job
contains more stuff such as the build/tmp directory on failure. This may
require restarting the job a few times until it randomly fails (which it
does maybe 1/5th of the time).