Added HAVE_INT128 test and variable to bitcoin-config.h
In the future ABLA implementation, we will need to use the __int128 type
(or emulate it if it's missing). As such, we need to add a cmake
configure test for it.

A test for __int128 has been added to src/config/CMakeLists.txt and this
variable will appear as defined to 0 or 1 in the generated
./src/config/bitcoin-config.h.

Test Plan
---

On your (presumably 64-bit) machine, regenerate the cmake configuration:

- `rm -fr build/ && mkdir build`
- `cd build`
- `cmake -GNinja ..`

Next, verify that ./src/config/bitcoin-config.h contains the define:

#define HAVE_INT128 1 (uncommented).

(Optional) Do the above on a machine that lacks the __i128 type (such as
ARM-32)... and verify that the define appears as #define HAVE_INT128 0
in ./src/config/bitcoin-config.h.