Compiler warning fix: Remove unused variable in policy.cpp
After commit dbbc51d9dbe434e57614b3e2f4e5ca364e295647 , the variable
`nDataOut` was no longer being used (read from). It was only being
written-to because previous to the above-mentioned commit it was used to
enforce relay policy.  But after the above commit, it stopped being
read-from.

On my latest clang it was producing a compile-time warning. This commit
fixes that.