Update regex'es in makeseeds.py ahead of May 2024 upgrade
Summary
---

This is a preliminary update of the regular expression patterns
in `makeseeds.py` to cover the nodes/versions in consensus
prior to the May 2024 upgrade.

It will still need to be updated as this information changes,
but this include BCHN v27 in the release, and adds recent
Knuth releases.

Test Plan
---

- Review