seeder: Don't use std::rand(); it's not guaranteed thread-safe
This commit fixes a potential UB issue with the seeder which was
accessing std::rand() from multiple threads. According to POSIX.1 this
function is not guaranteed to be thread-safe (it modifies internal
static state possibly without a lock held).  On gcc's C library it
happens to be thread-safe but this is not guaranteed for all platforms
and compilers.

So, to avoid potential UB here, we switched all usages of std::rand() to
using the app's own better rng: FastRandomContext.