seeder: Refactor C-style polymorphism -> C++ for CDnsThread
This commit refactors the dns.h utility functions to no longer do
C-style polymorphism with C-style callbacks and instead we do real C++
polymorphism here.  This avoids potential brittleness in the code (we
were passing around void * pointers to C-style callbacks which is not
ideal and is bugprone).

Also in this commit: avoid some UB by using a std::atomic for the
`nRequests` counter that is being accessed from multiple threads.