seeder: Fix bug where if DNS server can't start, app is silent with no indication of error
There was a bug in the seeder where if the DNS server failed to start
up, it was impossible to detect the situation. No errors were being
reported and the app would continue to run as if nothing was the matter.

This commit fixes the situation such that if the DNS server fails to
start, an error message is printed and the app aborts.

To test this new mechanism: pass a port that is already bound on the
machine or that is priviledged (such as port 53) via the -port=<N>
option. Seeder should complain and exit. Previous to this commit it
would silently ignore the error even though no DNS server was running.