seeder: Fix potential C++ UB by avoiding passing pointers through void *
Unlike in C, using void * in this way isn't guaranteed to work in C++ due
to strict aliasing rules. Instead, we create a tiny struct called "CrawlerArg"
which we stuff into the raw bytes of a void * pointer, and we unpack it
inside ThreadCrawler using std::memcpy. While this may seem strange, it
at least is 100% guaranteed to not produce any UB on any implementation.