linearize: Add the new testnets and rename diskmagic
Summary
---

The linearize example configuration did not have example values for
testnet4, scalenet and chipnet, this adds those.

Also the `netmagic` name is incorrect, as what is actually expected is
the disk magic (`diskMagic` in chain parameters), so we rename it to
`diskmagic`.

This also sets the example `max_height` to 823112, which is our last
checkpoint in `chainparams.cpp`.

Test plan
---

* `cp example-linearize.cfg linearize.cfg`
* make adjustments for your environment to `linearize.cfg`, make sure
  the `input` folder and the RPC settings are correct.
* `./linearize-hashes.py linearize.cfg > hashlist.txt`
* `./linearize-data.py linearize.cfg`
* for mainnet height 823112: `echo 4bac85da7b7666b6a9ac042e8dab63f5de2d6c41834bf70adc9e3e2e6c5441cd  bootstrap.dat | sha256sum -c`
* for testnet4 height 195959: `echo 570222234b7239067a9fe96ffe38d4543b5ce1f910b1f423a61ccd9ca0b5fe8d  bootstrap.dat | sha256sum -c`
* for chipnet height 195951: `echo a6606aa04f71f95a44b559e0c933212388961fec4f37c55159a8e494364a55e9  bootstrap.dat | sha256sum -c`