Fix socket leaks leading to process running out of file descriptors
If a peer node sends us garbage messages, CSeederNode::Run() throws an
ios_failure, which is fine. However, in this execution path the socket
ends up not being closed, so the process is leaking file descriptors in
this situation.

This led to a crash on my seeder after many months of operation.