net_processing: Avoid reading the block for MSG_FILTERED_BLOCK if no filter
Summary
---

The p2p message `MSG_FILTERED_BLOCK` is used to retrieve a partial "merkle block"
when a txn bloom filter is defined for the peer.  As per the spec, if there is
no filter defined, nothing is to be returned to the peer.

So far so good, right?

However, the extant code was reading the whole block into memory
regardless (deserializing it, etc), only to check if a filter is defined later on
in the function and then doing nothing if that is the case. So reading
the block in this case would be useless and waste cycles.

This nit just modifies the code to only really read the block into
memory if it's clear we will need it to build the merkle block -- that
is, if a filter is defined for the peer.

Test Plan
---

- Review. There are no behavioral changes, this is a pure code
  quality/slight performance nit.
- `ninja check && test/functional/test_runner p2p_filter`