bench: nit: Don't double-copy in VerifyScript.* benchmarks
Summary
---

This is a small nit in the `VerifyScripts.*` benchmarks that live in
`bench/verify_script.cpp.`

The benchmarks were loading block and coin data into memory twice, because
they were using `CDataStream` when unserializing the raw binary data. Instead,
we prefer to use the newer `VectorReader` which does in-place reads of
vectors (rather than copying the data to a private vector, as
`CDataStream` does).

This is a tiny code quality nit and does not change the behavior of the
benchmarks at all (nor should it even change their execution time). It
merely optimizes the benchmark setup code to not waste cycles and memory
on setup.

Test Plan
---

`ninja bench_bitcoin && src/bench/bench_bitcoin -filter='VerifyScript.*'`