LibAuth tests: Rename/refactor to support dropping-in of non-CHIP tests & benchmarks
For the upcoming work on the Targeted Virtual Machines Limits CHIP, we
will need to drop-in some benchmark tests from libauth, as well as more
regression tests. Some of the test and benches we plan to add don't fit neatly
into the "CHIPs" paradigm -- they are just tests and benchmarks unrelated
to any particular CHIP or feature being toggled on/off.

As such, we want to make the libauth-based tests more "generic" and not
just oriented towards CHIPs.

This commit does just that, moving the files around, renaming some
things, and allowing one to drop-in `.json` files anywhere under the
src/test/data/libuath_test_vectors/ dir, not just within `CHIPs`.

Also, some data structures were renamed to be more generic sounding, and
the `expected_test_fail_reasons.json` file was renamed to
`libauth_expectedtest_fail_reasons.json`.  The unit test
`libauth_chip_tests` was renamed to the more generic `libauth_tests`.

Documentation was also updated.

More detail: The existing concept of the "chip" or "chipName" that the
libauth facility was testing has been renamed to a "test pack" or
"packName".  Some types of test packs, such as CHIPs, have the
"ActivateFeature()" toggle where they can be toggled on/off. More
generic regression tests (in a future MR), will not be toggling anything
on/off since they will be testing existing VM behavior.

Also in this commit: `libauth_testing_setup.cpp` was making my IDE very unhappy
due to the huge "header" it was including, and I split up the
compilation/generation of the JSON files to a .h and a .cpp file to be
generated by the `generate_chip_test_vectors.py` script.