Switch ABLA (upgrade10) to use height-based activation, rather than MTP
Summary
---

After !1849, it is now safe to do the faster height-based activation
checks for ABLA (upgrade10) rather than the slower-to-calculate MTP
checks. This transition to height-based checking is what we typically do
in this codebase after an upgrade has been activated and checkpointed.

Updated all unit and functional tests to use the new height-based CLI arg
and checks.

Also in this MR: The "upgrade 10 activation block tracker" has now been
updated to track the upcoming tentative upgrade11 instead.

Test Plan
---

- Review
- Make sure the hard-coded heights for each chain for upgrade10
  are the first heights for which the MTP time >= activation time (`1715774400`
  for main, test3, test4, or `1700049600` for chipnet).
  - If you do use `getblockheader` to verify this, also these are the first heights
    where the verbose results from `getblockheader` returns some abla
    state associated with these blocks.
- `ninja all check-all`