trivial: Fix up the header util/heapoptional.h includes
Summary
---

This header was including <algorithm> needlessly (which is a large
header) and was not including <functional>, which it actually is using.

Also in this commit: a whitespace fix

Test Plan
---

No behavioral or other changes are introduced by this commit. It is
purely a trivial fix. However, for belt-and-suspenders:

- `ninja all check`