Imported latest LibAuth regression tests for the "2023" VM.
Imported latest tests from: https://github.com/bitjson/libauth/tree/d0d54b99445af46176804ec9a43c18a1edb9f139/src/lib/vmb-tests/generated

This required us to add a new concept to the libauth testing setup, the "scriptonly"
mode, which is a manual override done on a per-test basis.

This was required for some of the tests that do crazy things like spend more
than 21 million BCH or set locktime in the txn to high values. In "scriptonly" mode
we just evaluate the script for the designated input, after doing minimal standardness
checks, rather than run the txn through the full ATMP code path, so as
to allow for the small number of these unorthodox tests to make it to
the script VM for evaluation.