Put test_bitcoin common files into a library, use with bench_bitcoin
This allows us to use the test_bitcoin common files and data in
bench_bitcoin app.  For example, we will need to "see" the libauth tests
in bench_bitcoin for upcoming work.

This actually handles an old TODO from src/bench/CMakeLists.txt, so it
should be a welcome change.