bench: Support printing of supplemental stats; use this mechanism for LibAuth benches
This modifies the bench_bitcoin internal API to support benchmark
"categories" that can each optionally print a table at the end of
"supplemental" stats. This mechanism is used for the LibAuth benches to
support printing additional stats.

So now, if bench_bitcoin executes the LibAuth benches, it will print an
additional stats table specific to these benches at the end.