[backport] net processing: Only send a getheaders for one block in an INV
Summary
---

This is a backport of [Core#18962](https://github.com/bitcoin/bitcoin/pull/18962)

Original PR Message:

Headers-first is the primary method of announcement on the network. If a
node fell back sending blocks by inv, it's probably for a re-org. The
final block hash provided should be the highest, so send a getheaders
and then fetch the blocks we need to catch up.

Sending many GETHEADERS messages to the peer would cause them to send a
large number of potentially large HEADERS messages with redundant data, which
is a waste of bandwidth.

Test Plan
---

- `ninja all check-all`
- (Optional) Run a node, verify nothing broke. Note that this code path
  is almost never taken and only could be taken on a peer reorging. In
  that case the assumptions of this change hold and the latest hash is
  requested.