[backport] Add "overload' functional test + disable the penalty for PF_RELAY peers
Co-authored-by: Calin Culianu <calin.culianu@gmail.com>

Based on commit #5 of: https://github.com/bitcoin/bitcoin/pull/19988
Original commit: https://github.com/bitcoin/bitcoin/pull/19988/commits/de11b0a4eff20da3e3ca52dc90948b5253d329c5

Note we modified this commit:

We adapted this commit to *not* change the 100,000 tx inv tracking limit down
to 5,000, as Core did. We want to keep this backport strictly to not
modifying existing behavior, and existing code allowed a 100k limit, so
we keep that limit.

Also, the 5k limit that Core is now using seems low and perhaps is more
compatible with their chain, but for our chain, with 32x larger blocks,
we should keep the limit at 100k as it was before.