lint: Add check for usage of strerror in the codebase
Since strerror() is not thread-safe, we need to add a check to the
linter for usgae of this function.

Some exclusions are permitted for false positives in the regexp that
searches for it, but otherwise bare use of this function should be
forbidden in the codebase now, owing to this lint check.