Fix compiler error in blockstorage.cpp for older compilers such as on ubuntu bionic
The compiler was unable to find the proper constructor for the
std::optional return value, so we explicitly basically construct the
return object. RVO should elide any copies in doing it this way. Code is
equivalent to previous just is more likely to compile on an older GCC.

See this error when building ubuntu ppas:
https://launchpadlibrarian.net/739129199/buildlog_ubuntu-bionic-amd64.bitcoin-cash-node_27.1.0-bionic1_BUILDING.txt.gz