RPC: Add input tokenData and scriptPubKey to getrawtransaction verbosity=2
We modify the `getrawtransaction` RPC method when verbosity=2 to add
more information about the inputs. Namely, the `scriptPubKey` of the
prevout, as well as the `tokenData` of the prevout (optional).

This closes #527, and makes getrawtransaction with verbosity=2 behave
like the verbose `getblock`, whereby we fully show the input's scriptPubKey
and tokenData (if token input).

Funtional tests were updated to test for the new keys.