backport: Fix compile warning, fix UB, and optimize CoinUndoSpend
Summary
---

This is a backport of https://reviews.bitcoinabc.org/D14010. It fixes a
compile warning which is mis-use of std::move on a const object.

However, actually the code in question doing some UB (`const_cast`
there to write to a const value! Yikes!).  Luckily, it is harmless but to
future-proof this codebase it is important to fix this.

As a side effect this commit also ends up being a performance
optimization in some cases.

Note that it's safe to use move here because all code paths basically
are working with ephemeral BlockUndo objects anyway and one can review
and see that it is the case that these changes are safe.

Original author message:

    This is the consequence of a missed commit from core#10195:
    https://github.com/bitcoin/bitcoin/pull/10195/commits/bd83111a0fcfdb97204a0180bcf861d3b53bb6c2

Test Plan
---

- `ninja all check`