Nuke lint-cpp-void-parameters
Summary
---

This linter has a nonzero chance of returning false positives since it
doesn't really understand C++ syntax, it just uses a simple regex.

It is also my opinion that there is no real "harm" in C-style `int foo(void);`
declarations anyway, so it's not worth writing the code to make this
linter "robust".

Test Plan
---

- `ninja check-lint`