Upgrade12: Update activation_tests, test_bitcoin, and the g_upgrade*tracker
This commit updates test_bitcoin to accept -upgrade12activationtime= as
well as updates the activation_tests to test the new IsUpgrade12Enabled
function.

Also in this commit: bump the "upgrade height tracker" in validation.cpp
to pay attention to upgrade12 rather than upgrade11. This is safe to do
since it turns out that upgrade11 doesn't need to track the exact
activation height (unlike the previous upgrades such as cashtokens and
abla which did need to track this).